Add option to supply classpath.txt
