+++ CHANGELOG.md
@@ -5,6 +5,8 @@ We use [semantic versioning][semver]
 - [breaking change] `--ignore-duplicates` (and `-d` option in convert tool) have been replaced with `--duplicates` option
 - [breaking change] `--filter` option in convert tool has been renamed to `--includes`
 - [breaking change] `--exclude` option in convert tool has been renamed to `--excludes`
+- [breaking change] `--classDir` option in convert tool has been renamed to `--class-dir`
+- [feature] `--class-dir` option allows to pass in a *.txt file with the class directories/jars
 
 # 15.5.0
 - [feature] add TIA client library for integrating TIA in your custom test framework
