+++ agent/README.md
@@ -76,8 +76,8 @@ patterns with `*`, `**` and `?`.
 
 - `class-dir`: the path under which all class files of the profiled application are stored. Normally, this is inferred
   by the agent automatically. For some application, profiling performance may improve if you specify it explicitly. May be
-  a directory or a Jar/War/Ear/... file. Separate multiple paths with a semicolon. (For details see path format section 
-  above)
+  a directory or a Jar/War/Ear/... file. Separate multiple paths with a semicolon. You may also supply a *.txt 
+  file with classpath entries one per line (For details see path format section above)
 - `interval`: the interval in minutes between dumps of the current coverage to an XML file (Default is 60). If set to 
   0 coverage is only dumped at JVM shutdown.
 - `dump-on-exit`: whether a coverage report should be written on JVM shutdown (Default is true).
